feat: watchdog (#10)
Watchdog is implemented to handle a worker failure.
A tcpstore is used to check the liveness of worker.
Upon detecting at least one worker's failure, the watchdog
lets the world manager know the broken world. The world manager then
cleans up its state on the broken world and BrokenWorldException is
raised. Caveat: while a failure is handled, data transfer is
temporarily halted. This needs further investigation.