patch: pytorch v2.2.1 patch file for multiworld
To support parallel init_process_group() invocations, pytorch's
distributed_c10d.py is patched. Instead of a singleton, _world, we
define _worlds as a dictionary so that different World instance
can be maintained. In order to switch between different worlds,
we introduce a new variable, "name" in many of functions in the file.

The changes introduced in distributed_c10d.py lets modifications in
world_manager.py and world_communicator.py. Unnecessary code was
cleaned up.