nit: keyerror exception handling
There is a chance that remove_world in world_manager is called twice.
The funcall call removes a key from worlds_stores.
So, the second call leads to KeyError.
We currently can't ensure call the function only once because
the current remove_world is not clean due to a deadlock issue.
So, we mask the key error.