refactor: nccl's async error handling in pytorch (#59)
The default option to handle errors of nccl operations in PyTorch is to tear
down the process. In a multiworld setting, we need to handle error gracefully
by tearing down the world that encounter the errors. In order to do so, we set
the environment variable, TORCH_NCCL_ASYNC_ERROR_HANDLING, with "2"
(i.e., CleanupOnly) in the World Manager''s init function.
In this way, users don't need to set the environment variable in their
applications. Thus, in the examples, we removed the code snippet that sets
the environment variable.