feat: boolean function to check if world is broken
A broken world exception is raised for collective operations such as
send and recv. However, if those operations are not called and
runtime error is not raised, the broken world won't be noticed until
any of the operations is called. To check if the world is broken or not
without calling the operations, a new function is introduced.