fix: watchdog deadlock check on recovery
When a broken world is identified, watchdog starts to send SIGUSR1 messages. But whenever these worlds are recovered, we need to bock the thread execution for the previous worlds. So whenever a new set of broken worlds is received, the event is unblocked and it can continue to send messages. The cleanup of the broken worlds is already handled by watchdog.