misc: clean watchdog logs (#107)
Removed logs that are misleading from watchdog. Whenever watchdog identifies a broken world, it starts to send SIGUSR1 messages to check the deadlock. These messages are sent continuously for 50 seconds. That is correct, but the log messages are misleading. If there is an actual deadlock, there are logs that will be displayed for that reason.